requirements.txt
tensorflow-cpu>=2.15,<3.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.8
